Move screening players into their own file; add a "Play" button to cover up play icon on video
